Completed Stewart Flags and Such Invoice v.1.0
Completed Stewart Flags and Such Invoice v.1.0, unzip and inside is the executable of the Invoice Generator as well as the Python Calculator. Invoice Generator was created in Tkinter and exports as a .docx file with a "save as" function to choose it's storage point. Invoice will also add the date and time to the end of the invoice information.